qa_project/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
docker==7.1.0
psycopg2-binary==2.9.11
pytest==9.0.1
pytest-django==4.11.1
